tests: Skip object read-only tests for PyPy
Skip the object read-only tests that are known to fail on PyPy. They
are not critical to the module serving its main purpose, and I believe
it would be better to let PyPy users know whether the core algorithms
work.

This can be amended with a specific PyPy version check once upstream
fixes https://bitbucket.org/pypy/pypy/issues/2686.

Closes: https://github.com/tiran/pysha3/issues/13